+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project are documented in this file.
 ## Unreleased
 
 ## v2.8.7 (2026-02-13)
+- **mirrors-fix** — For mirrors, always use mirror in imageContentSource template
+- **release-script-fix** — Fix `rg` dependency (use `grep`), push only new tag instead of all tags
 
 ## v2.8.6 (2026-02-13)
 - **fix-manifestsconfigmapref** — Move `manifestsConfigMapRef` from ClusterDeployment to AgentClusterInstall; `provisioning` and `clusterInstallRef` are mutually exclusive in the Hive CRD